2026-06-10 08:26 partyBoard.jsp 파티 신청 처리 및 댓글 추가 구현 , 댓글 삭제 처리 미구현

+++ src/main/java/com/noexit/app/model/PartyApplyDTO.java
@@ -13,7 +13,7 @@ public class PartyApplyDTO
 	private String nickName;
 	private String gender;
 	private int age;
-	private int temp;
+	private double temp;
 	private String applyComment;
 	private String createdAt;
 }
